export function instead of object
